+++ src/components/incidents/incidentForms/LostAndFoundForms/steps/Step1Info.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Step1Info = () => {
+    return (
+        <div>Step1Info</div>
+    )
+}
+
+export default Step1Info
