+++ src/components/incidents/incidentForms/LostAndFoundForms/steps/Step2Actions.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Step2Actions = () => {
+    return (
+        <div>Step2Actions</div>
+    )
+}
+
+export default Step2Actions
